Pulled validators into their own file
