+++ routes/validators.js
@@ -1,0 +1,60 @@
+const { check, validationResult } = require("express-validator");
+const db = require("../db/models");
+
+const signupValidators = [
+  check("userName")
+    .exists({ checkFalsy: true })
+    .withMessage("Please provide a value for User Name")
+    .isLength({ max: 250 })
+    .withMessage("User Name must not be more than 150 characters long"),
+  check("email")
+    .exists({ checkFalsy: true })
+    .withMessage("Please provide a value for Email Address")
+    .isLength({ max: 250 })
+    .withMessage("Email Address must not be more than 255 characters long")
+    .isEmail()
+    .withMessage("Email Address is not a valid email")
+    .custom((value) => {
+      return db.User.findOne({ where: { emailAddress: value } }).then(
+        (user) => {
+          if (user) {
+            return Promise.reject(
+              "The provided Email Address is already in use by another account"
+            );
+          }
+        }
+      );
+    }),
+  check("password")
+    .exists({ checkFalsy: true })
+    .withMessage("Please provide a value for Password")
+    .isLength({ max: 50 })
+    .withMessage("Password must not be more than 50 characters long"),
+  check("confirmPassword")
+    .exists({ checkFalsy: true })
+    .withMessage("Please provide a value for Confirm Password")
+    .isLength({ max: 50 })
+    .withMessage("Confirm Password must not be more than 50 characters long")
+    .custom((value, { req }) => {
+      if (value !== req.body.password) {
+        throw new Error("Confirm Password does not match Password");
+      }
+      return true;
+    }),
+];
+
+const loginValidators = [
+  check("emailAddress")
+    .exists({ checkFalsy: true })
+    .withMessage("Please provide a value for Email Address"),
+  check("password")
+    .exists({ checkFalsy: true })
+    .withMessage("Please provide a value for Password"),
+];
+
+module.exports = {
+  db,
+  validationResult,
+  loginValidators,
+  signupValidators,
+};
